restructure codeadapter and fixed an error in the META-INF in SDM Compiler Democles
