+++ org.moflon.gt.mosl.codeadapter/META-INF/MANIFEST.MF
@@ -12,5 +12,4 @@ Require-Bundle: org.eclipse.emf.ecore;bundle-version="2.12.0",
  org.eclipse.equinox.registry,
  org.eclipse.core.resources
 Bundle-Vendor: TU Darmstadt
-Export-Package: org.moflon.gt.mosl.codeadapter,
- org.moflon.gt.mosl.codeadapter.codeadapter
+Export-Package: org.moflon.gt.mosl.codeadapter
