+++ org.moflon.gt.mosl.codeadapter/src/org/moflon/gt/mosl/codeadapter/ObjectVariableBuilder.java
@@ -1,4 +1,4 @@
-package org.moflon.gt.mosl.codeadapter.codeadapter;
+package org.moflon.gt.mosl.codeadapter;
 
 import java.util.LinkedList;
 import java.util.List;
