+++ org.moflon.gt.mosl.codeadapter/src/org/moflon/gt/mosl/codeadapter/utils/PatternKind.java
@@ -1,0 +1,26 @@
+package org.moflon.gt.mosl.codeadapter.utils;
+
+public enum PatternKind {
+   BLACK, GREEN, RED, EXPRESSION, BINDING_AND_BLACK, BINDING;
+
+   public String getSuffix()
+   {
+      switch (this)
+      {
+      case BINDING_AND_BLACK:
+         return "bindingAndBlack";
+      case BLACK:
+         return "black";
+      case EXPRESSION:
+         return "expression";
+      case GREEN:
+         return "green";
+      case RED:
+         return "red";
+      case BINDING:
+         return "binding";
+      default:
+         return null;
+      }
+   }
+}
